Initial port to netcore2.1
