--- StatsDHelper.Tests/packages.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="FluentAssertions" version="3.3.0" targetFramework="net45" />
-  <package id="NUnit" version="2.6.4" targetFramework="net45" />
-  <package id="RhinoMocks" version="3.6.1" targetFramework="net45" />
-</packages>
