+++ StatsDHelper.sln
@@ -1,10 +1,8 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2012
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StatsDHelper", "StatsDHelper\StatsDHelper.csproj", "{299064CA-B8A7-4DBB-B93E-B912618AD828}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StatsDHelper.Tests", "StatsDHelper.Tests\StatsDHelper.Tests.csproj", "{6F18BABE-CC5D-494F-AEF8-24F465E63EA6}"
-EndProject
+# Visual Studio 15
+VisualStudioVersion = 15.0.27703.2042
+MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{6D7286E0-BF11-4718-A7DA-0C2F2591B1DD}"
 	ProjectSection(SolutionItems) = preProject
 		.nuget\NuGet.Config = .nuget\NuGet.Config
@@ -19,27 +17,39 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		README.md = README.md
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StatsDHelper", "StatsDHelper\StatsDHelper.csproj", "{AFA0B443-3740-48C5-A3D1-A7C340E915D8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StatsDHelper.Tests", "StatsDHelper.Tests\StatsDHelper.Tests.csproj", "{352303DD-9D03-473E-819C-6440821E6C3D}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release 4.0|Any CPU = Release 4.0|Any CPU
 		Release 4.5|Any CPU = Release 4.5|Any CPU
+		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{299064CA-B8A7-4DBB-B93E-B912618AD828}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{299064CA-B8A7-4DBB-B93E-B912618AD828}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{299064CA-B8A7-4DBB-B93E-B912618AD828}.Release 4.0|Any CPU.ActiveCfg = Release 4.0|Any CPU
-		{299064CA-B8A7-4DBB-B93E-B912618AD828}.Release 4.0|Any CPU.Build.0 = Release 4.0|Any CPU
-		{299064CA-B8A7-4DBB-B93E-B912618AD828}.Release 4.5|Any CPU.ActiveCfg = Release 4.5|Any CPU
-		{299064CA-B8A7-4DBB-B93E-B912618AD828}.Release 4.5|Any CPU.Build.0 = Release 4.5|Any CPU
-		{6F18BABE-CC5D-494F-AEF8-24F465E63EA6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6F18BABE-CC5D-494F-AEF8-24F465E63EA6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6F18BABE-CC5D-494F-AEF8-24F465E63EA6}.Release 4.0|Any CPU.ActiveCfg = Release|Any CPU
-		{6F18BABE-CC5D-494F-AEF8-24F465E63EA6}.Release 4.0|Any CPU.Build.0 = Release|Any CPU
-		{6F18BABE-CC5D-494F-AEF8-24F465E63EA6}.Release 4.5|Any CPU.ActiveCfg = Release|Any CPU
-		{6F18BABE-CC5D-494F-AEF8-24F465E63EA6}.Release 4.5|Any CPU.Build.0 = Release|Any CPU
+		{AFA0B443-3740-48C5-A3D1-A7C340E915D8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AFA0B443-3740-48C5-A3D1-A7C340E915D8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AFA0B443-3740-48C5-A3D1-A7C340E915D8}.Release 4.0|Any CPU.ActiveCfg = Release|Any CPU
+		{AFA0B443-3740-48C5-A3D1-A7C340E915D8}.Release 4.0|Any CPU.Build.0 = Release|Any CPU
+		{AFA0B443-3740-48C5-A3D1-A7C340E915D8}.Release 4.5|Any CPU.ActiveCfg = Release|Any CPU
+		{AFA0B443-3740-48C5-A3D1-A7C340E915D8}.Release 4.5|Any CPU.Build.0 = Release|Any CPU
+		{AFA0B443-3740-48C5-A3D1-A7C340E915D8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AFA0B443-3740-48C5-A3D1-A7C340E915D8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{352303DD-9D03-473E-819C-6440821E6C3D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{352303DD-9D03-473E-819C-6440821E6C3D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{352303DD-9D03-473E-819C-6440821E6C3D}.Release 4.0|Any CPU.ActiveCfg = Release|Any CPU
+		{352303DD-9D03-473E-819C-6440821E6C3D}.Release 4.0|Any CPU.Build.0 = Release|Any CPU
+		{352303DD-9D03-473E-819C-6440821E6C3D}.Release 4.5|Any CPU.ActiveCfg = Release|Any CPU
+		{352303DD-9D03-473E-819C-6440821E6C3D}.Release 4.5|Any CPU.Build.0 = Release|Any CPU
+		{352303DD-9D03-473E-819C-6440821E6C3D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{352303DD-9D03-473E-819C-6440821E6C3D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {53872B55-E85B-466A-A0DC-594ED9812E94}
 	EndGlobalSection
 EndGlobal
