--- StatsDHelper/StatsDHelper.nuspec
@@ -1,23 +1,0 @@
-﻿<?xml version="1.0"?>
-<package >
-  <metadata>
-    <id>statsd-helper</id>
-    <version>$version$</version>
-    <authors>PayByPhone</authors>
-    <projectUrl>https://github.com/paybyphone/statsd-helper</projectUrl>
-    <licenseUrl>http://opensource.org/licenses/Apache-2.0</licenseUrl>
-    <owners>PayByPhone</owners>
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <description>Simple wrapper around statsd-csharp-client which prepends the metric name with the hosts Fully Qualified Domain Name and a configurable ApplicationName prefix.</description>
-    <releaseNotes>Initial Release</releaseNotes>
-    <copyright>Copyright 2015</copyright>
-    <tags>statsd statsd-csharp-client prefix fqdn</tags>
-    <dependencies>
-      <dependency id="StatsdCsharpClient" version="[1.3.0,)" />
-    </dependencies>
-  </metadata>
-  <files>
-    <file src="bin\Release\net40\StatsDHelper.dll" target="lib\net40\StatsDHelper.dll" />
-    <file src="bin\Release\net45\StatsDHelper.dll" target="lib\net45\StatsDHelper.dll" />
-  </files>
-</package>
