+++ StatsDHelper/StatsDHelperConfig.cs
@@ -1,0 +1,9 @@
+﻿namespace StatsDHelper
+{
+    public class StatsDHelperConfig
+    {
+        public string StatsDServerHost { get; set; }
+        public int? StatsDServerPort { get; set; }
+        public string ApplicationName { get; set; }
+    }
+}
