+++ StatsDHelperFramework.Tests/App.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<configuration>
+  <appSettings />
+</configuration>
